Try to facter out into parts
